requirements.txt
langchain_groq==0.2.4
python-dotenv==1.0.1
streamlit==1.41.1
